Git is now a protocol.
This will allow it to be backed by a different implemenation.
